move everything into MainApp.tsx just so can use statusBarHeight
