+++ screens/NativeBasePOC.tsx
@@ -3,7 +3,7 @@ import { StyleSheet, Image } from 'react-native';
 
 import { Container, Content, Card, CardItem, Thumbnail, Text, Button, Icon, Left, Body, CheckBox } from 'native-base';
 import { DrawerNavigationProp } from '@react-navigation/drawer';
-import { DrawerParamList } from '../App'
+import { DrawerParamList } from './MainApp'
 
 type NativeBasePOCProps = {
   navigation: DrawerNavigationProp<DrawerParamList, 'Home'>;
